refactor : extract popover from userItem component
